Container | XYContainer | Auto Margin: Set scales before updating domain
Fixes the problem with axis tick labels going off screen because the auto margins were not calculated properly.

It was happening when a different scale (e.g. Scale.scaleTime()) is provided to the container (and it generates ticks differently from the default scale), because this new scale was not passed to the Axis component during pre-render