Container | XY: Handle edge cases when `scaleByDomain` is `true`
There could be situations when there are no data points, or only a single data point remains after `filterDataByRange` when calculating the domain. This was causing XY charts to go off-screen because the domain was calculated improperly.

This change addresses the issue by including neighboring data points in the filtered results.